[toplevelname] switching to int64 to avoid casts later in the pipeline

+++ model/stats.go
@@ -28,7 +28,7 @@ type Count struct {
 	Measure string `json:"measure"` // represents the entity we count, e.g. "hits", "errors", "time" (was Name)
 	TagSet  TagSet `json:"tagset"`  // set of tags for which we account this Distribution
 
-	TopLevel int `json:"top_level"` // number of top-level spans contributing to this count
+	TopLevel int64 `json:"top_level"` // number of top-level spans contributing to this count
 
 	Value float64 `json:"value"` // accumulated values
 }
@@ -40,7 +40,7 @@ type Distribution struct {
 	Measure string `json:"measure"` // represents the entity we count, e.g. "hits", "errors", "time"
 	TagSet  TagSet `json:"tagset"`  // set of tags for which we account this Distribution
 
-	TopLevel int `json:"top_level"` // number of top-level spans contributing to this count
+	TopLevel int64 `json:"top_level"` // number of top-level spans contributing to this count
 
 	Summary *quantile.SliceSummary `json:"summary"` // actual representation of data
 }
